+++ Cmdr/BuiltInCommands/Debug/uptime.lua
@@ -1,0 +1,7 @@
+return {
+	Name = "uptime";
+	Aliases = {};
+	Description = "Returns the amount of time the server has been running.";
+	Group = "DefaultDebug";
+	Args = {};
+}